Added new CollectedNote comment for auton
Added a new command 'CollectedNote' with a timer feature and linked
it to an Indexer subsystem. In the Indexer subsystem, updated
'hasShot' method to return the 'didShoot' status,
added the 'clearHasShot' method to reset 'didShoot' status,
and added 'hasNote' method to check the current PieceState.

Also, made improvements to the periodic method to update
'didShoot' status and set PieceState.